embed and backend overhaul
made a lot of edits to pytube to get the lengths and the views of videos working, but now the database stores more information about the requested song, the search shows views and length, and the embed was fixed accordingly

+++ bot.py
@@ -16,9 +16,12 @@ from pytube import Search
 from util import *
 import re
 from discord.ext.commands import cooldown, BucketType
+import pafy
+from debug import debug
 
 load_dotenv()
 TOKEN = os.getenv('TOKEN')
+MAXVIDEOLENGTH = 3600
 bot = commands.Bot(command_prefix='!', intents = discord.Intents.all())
 myID = 281911981089226762
 
@@ -26,6 +29,8 @@ myID = 281911981089226762
 @bot.event
 async def on_ready():
     print(f'{bot.user} has connected to Discord!')
+    # await bot.change_presence(status=discord.Status.dnd, activity=discord.Activity(type=discord.ActivityType.watching, name="Tissi Coding"))
+    await bot.change_presence(status=discord.Status.dnd,activity=discord.CustomActivity(name='david sucks ass' ,emoji='🖥️'))
     try:
         synced = await bot.tree.sync()
         print(f"Synced {len(synced)} command(s)")
@@ -61,9 +66,12 @@ async def hello(interaction: discord.Interaction):
 @bot.tree.command(name = "say", description='Repeats what you say')
 @app_commands.describe(thing_to_say = "what should i say")
 @app_commands.describe(secret = "is this a secret?")
-async def say(interaction: discord.Interaction, thing_to_say: str, secret: bool = False):
+async def say(interaction: discord.Interaction, thing_to_say: str, secret: bool = True):
     print(f'Command: {interaction.command.name} was invoked by {interaction.user.nick if interaction.user.nick is not None else interaction.user.global_name} in {interaction.guild.name} - {interaction.channel.name}')
-    await interaction.response.send_message(f"{interaction.user.mention} said: {thing_to_say}" if not secret else f"{thing_to_say}")
+    if not secret:
+        await interaction.response.send_message(f"{interaction.user.mention} said: {thing_to_say}", ephemeral = True, delete_after=5)
+    else:
+        await interaction.channel.send(f"{thing_to_say}")
     
 @bot.tree.command(name = "tierlist", description='Starts voting for the Teir List')
 @app_commands.describe(timer = "is this timed? (true: timer, false: command to end)")
@@ -77,7 +85,7 @@ async def tierlist(interaction: discord.Interaction, timer: bool = False, voting
     User = Query()
     res = db.search(User.channel == interaction.channel.id)
     if len(res) >= 1:
-        await interaction.response.send_message(f"already one tierlist ongoing", ephemeral = True, delete_after=5)
+        await interaction.response.send_message(f"already one tierlist ongoing, try /endtierlist", ephemeral = True, delete_after=7)
         return
     
     await interaction.response.send_message(f"kys!!!!", ephemeral = True, delete_after=3)
@@ -86,6 +94,7 @@ async def tierlist(interaction: discord.Interaction, timer: bool = False, voting
     
     members = [x for x in interaction.channel.members if (not x.bot and x.id != 704872368572465163) or x.id == 1186185707258134560]
     vote_msg_list = []
+    await interaction.channel.send('@everyone starting tierlist')
     for member in members:
         vote_msg = await interaction.channel.send(f'Where should {member.global_name if member.global_name is not None else member.name} be placed?')
         await vote_msg.add_reaction('🆘')
@@ -122,7 +131,7 @@ async def endtierlist(interaction: discord.Interaction):
     res = db.search(User.channel == interaction.channel.id)
     
     if len(res) == 0 or (res[0]['channel'] == res[0]['vote_msg_list'] and res[0]['channel'] == res[0]['memberids']):
-        await interaction.response.send_message(f"no tierlist to end", ephemeral = True, delete_after=5)
+        await interaction.response.send_message(f"no tierlist to end, try /tierlist", ephemeral = True, delete_after=7)
         return
     
     await interaction.response.send_message(f"KYS!!!!", ephemeral = True, delete_after=3)
@@ -131,7 +140,6 @@ async def endtierlist(interaction: discord.Interaction):
     vote_msg_list = res['vote_msg_list']
     memberids = [await bot.fetch_user(x) for x in res['memberids']]
     await createlist(interaction.channel, vote_msg_list, memberids)
-    
     # Remove the query from the database
     db.remove(User.channel == interaction.channel.id)
 
@@ -145,7 +153,11 @@ async def play(interaction: discord.Interaction, query: str):
     queue = TinyDB('queue.json')
     User = Query()
     res = queue.search(User.server == interaction.guild.id)
-    
+    if len(res) == 0:
+        queue.insert({'server': interaction.guild.id, 'queue': []})
+        User = Query()
+        res = queue.search(User.server == interaction.guild.id)
+        
     if (os.path.exists(f'vids/{interaction.guild.id}.webm') or os.path.exists(f'vids/{interaction.guild.id}_queue') or len(res[0]['queue']) >=1) and voice is None:
         if (os.path.exists(f'vids/{interaction.guild.id}.webm')): os.remove(f'vids/{interaction.guild.id}.webm')
         if (os.path.exists(f'vids/{interaction.guild.id}_queue')): shutil.rmtree(f'vids/{interaction.guild.id}_queue')
@@ -161,7 +173,11 @@ async def play(interaction: discord.Interaction, query: str):
         video_url = checker_url + (str(video_id.group(0)) if video_id is not None else "0")
         if video_id and requests.get(video_url).status_code == 200:
             query = f'https://www.youtube.com/watch?v={str(video_id.group(0))}'
-            addtoQueue(interaction,query)
+            videoObj = Search(query).results[0]
+            if videoObj.length_seconds > MAXVIDEOLENGTH:
+                await interaction.response.send_message(f"Video is too long, go fuck yourself.", ephemeral = True, delete_after=7)
+                return
+            addtoQueue(interaction,videoObj)
             if voice is None:
                 await interaction.response.send_message(f"playing", ephemeral = True, delete_after=3)
                 await downloadAndPlay(interaction,query,uservoice)
@@ -172,16 +188,35 @@ async def play(interaction: discord.Interaction, query: str):
             await interaction.response.send_message(f"invalid youtube link", ephemeral = True, delete_after=5)
         return
     
-    await interaction.response.send_message(f"Searching!", ephemeral = True, delete_after=3)
-    
     s = Search(query)
-    selectionmessage = [f"{k+1} : {v.title}\n" for k,v in enumerate(s.results[0:10])]
-    select = Select(options = [discord.SelectOption(label = f'{k+1}: {v.title}'[:99] if len(f'{k+1} - {v.title}') > 99 else f'{k+1} - {v.title}', description = f"{v.watch_url}", value=k) for k,v in enumerate(s.results[0:10])])
+    if (len(s.results) == 0):
+        await interaction.response.send_message(f"no results found", ephemeral = True, delete_after=5)
+        return
     
-    async def my_mycallback(ints):
+    await interaction.response.send_message(f"Searching!", ephemeral = True, delete_after=3)
+    # - {v.length} - {v.viewCountText}
+    selectionmessage = [f" {k+1} : {v.title}\n" if k!=9 else f"{k+1} : {v.title}\n" for k,v in enumerate(s.results[0:10])]
+    # print(s.results[0].views)
+    # print(s.results[0].length)
+    # debug(youtubeobj,"youtubeobj.json")
+    # video = pafy.new(s.results[0].watch_url)
+    # video.viewcount
+    # video.duration
+    #work on this, gn
+    select = Select(options = [discord.SelectOption(label = f'{k+1} : {v.title}'[:99] if len(f'{k+1}: {v.title}') > 99 else f'{k+1} : {v.title}', description = f'{v.length} - {v.viewCountText}', value=k) for k,v in enumerate(s.results[0:10])])
+    
+    async def my_mycallback(ints:Interaction[Client]):
+        if ints.user.id != interaction.user.id: 
+            await ints.response.send_message(f"You cant select this video, you did not search it", ephemeral = True, delete_after=7)
+            return
+        videoObj = s.results[int(select.values[0])]
+        if videoObj.length_seconds > MAXVIDEOLENGTH:
+            await ints.response.send_message(f"Video is too long, go fuck yourself.", ephemeral = True, delete_after=7)
+            return
         await newint.delete()
-        videourl = s.results[int(select.values[0])].watch_url
-        addtoQueue(interaction,videourl)
+        
+        videourl = videoObj.watch_url
+        addtoQueue(interaction,videoObj)
         #runqueue instead
         if voice is None:
             # await interaction.response.send_message(f"playing", ephemeral = True, delete_after=3)
@@ -193,7 +228,10 @@ async def play(interaction: discord.Interaction, query: str):
     select.callback = my_mycallback
     view = View()
     view.add_item(select)
-    newint = await interaction.channel.send(f'```'+''.join(map(str, selectionmessage))+'```',view=view)
+    # davidsid = 382271649724104705
+    #if interaction.user.id != davidsid else f'{interaction.user.mention} Searched: "gay furry porn"\n'
+    searchstring = f'{interaction.user.mention} Searched: "*{query}*"\n' 
+    newint = await interaction.channel.send(f'{searchstring}```'+''.join(map(str, selectionmessage))+'```',view=view)
 
 # @bot.tree.command(name = "stop", description='stops music?')
 # async def stop(interaction: discord.Interaction):
